🚚 Rename: boards 컴포넌트 폴더명 FreeBoards 로 변경
